chore: renamed job description to job details
